requirements.txt
# Requiments for running notebooks
-e .

jupyter
llama-index
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
ragas
python-dotenv
